refactor: move detecting language logic from RenderMarkdown to detectLanguage helper
