+++ src/components/FillCodeTask/index.tsx
@@ -3,6 +3,7 @@ import styles from './index.module.css'
 import AnimatedOption from '../AnimatedOption'
 import type { UserSelectAnswer } from '../../hooks/tasks/useUserTaskAnswer'
 import type { TaskAnswer, TaskWithAnswers } from '../../app/store/slices/taskSlice'
+import { getCleanCode } from '../../libs/helpers/detectLanguage.ts'
 
 interface FillCodeTaskProps {
   code: string
@@ -19,7 +20,8 @@ const FillCodeTask: React.FC<FillCodeTaskProps> = ({
   selectedOption,
   checked
 }) => {
-  const codeParts = code.split('PLACEHOLDER_ANSWER')
+  const cleanCode = getCleanCode(code);
+  const codeParts = cleanCode.split('PLACEHOLDER_ANSWER')
   const answerSlotRef = useRef<HTMLSpanElement>(null)
   const [animatingOption, setAnimatingOption] = useState<TaskAnswer | null>(null)
